Struct ipcrypt::State

pub struct State(_, _, _, _);

The inner state permutations are build on. Input and Output types are converted to an from this type.

You could provide custom From and Into implementations for local types, and then use encrypt and decrypt directly on those types.